tshwctl: Add RTC temp output gated behind define
Pass CFLAGS="-DRTCTEMP" to enable this output.

The ISL RTC is handled by the kernel driver but that
driver does not enable any of the temperature compensation.
Temperature compensation must be enabled to get the RTC temp.
Enabling temp. comp. will also start to trim the RTC itself. Thus,
the reason the RTC temp is gated behind a define is a precaution
against any issues/conflicts that might arise from the RTC
compensating itself while the system time is doing the same thing.

In the stock TS-7670 images, there is no kernel RTC driver and
the RTC is allowed to freely run. Only setting or getting the
system time at startup and shutdown.